Start reorganizing the Environment API
Construction of an Environment can now be done using the builder
pattern. Methods have been added to set the behavior when a file
read/write or a command execution is attempted.

The methods to pass an owned input/output stream to the environment have
been removed for now. It's not clear how these would be useful, and the
Program API doesn't expose this functionality anyway.